Move auth manual testcases
